bonus status messages added

+++ exer9/train-tickets/server/mongodb/actions/deleteTicket.js
@@ -5,7 +5,10 @@ export default async function deleteTicket(data) {
     try {
         await connectDB()
         const { identifier } = data
+        const exists = Ticket.exists({ _id: identifier })
+        if (!exists) return false
         await Ticket.findByIdAndDelete(identifier)
+        return true
     } catch (e) {
         console.log(e)
     }
